Refactor Transfers: split view and hook logic
